Replace sleep with try_n_times function in setup.sh
Instead of using sleep for 10 seconds in `setup.sh`, a function `try_n_times` was introduced. This function will try to run a command and match its output with an expected output. If it does not match, it waits for a bit and then retries. It repeats this process a specified number of times until it succeeds or reaches the max try limit. The function is then used to enhance the smoke test process.