feature(gateway): add local kwil gateway for testing (#149)
* Add local certificate and setup for dev-gateway

This commit introduces local certificate files for secure development. A new Taskfile.yml action 'setup:local-cert' has been added for trusting these local certificates. The dev-gateway has been configured for docker deployment by adding dev-gateway.dockerfile and dev-gateway-compose.yaml. A .gitignore file is updated to ignore the private binary.

* Add local gateway testing instructions

This commit adds a README to the dev-gateway deployment directory with instructions on how to test the Gateway locally. It includes guidelines about software requirements, steps for setup and a command to test the successful operation of the Gateway.

* fix incorrect conflicts

* fix incorrect conflicts

* Add Kwil Gateway section to README

A new section has been added to the README file explaining the optional use of the Kwil Gateway. It elaborates on the Kwil Gateway's function as a load-balancer with authentication capability for Proof of Authority Kwil blockchain networks. Instructions for local testing and development have also been included.

* Add start local gateway task to Taskfile

A new task, 'start:local-gateway', has been included in the Taskfile.yml. This task is not required for development but can be useful for testing the local gateway's configuration. The task will start the local gateway using a specific docker compose command.

---------

Co-authored-by: MicBun <michaelboentarman@gmail.com>