docs: define common terminology document (#170)
* Add TERMINOLOGY.md and update README

A new file `TERMINOLOGY.md` has been added that explains the project-specific terms. This document will help both developers and users understand the terminology used in the TSN-DB project. The `README.md` file was also updated to include a reference to the new terminology file.

* Update TERMINOLOGY.md

* Add definitions for "whitelist" and "private key" in TERMINOLOGY.md

Additional definitions for "whitelist" and "private key" have been added to the TERMINOLOGY.md document to improve the overall understanding. "Whitelist" pertains to the permissions a whitelisted wallet may have, while "private key" refers to a secret key associated with a wallet.

* Apply suggestions from code review

Co-authored-by: Vadim <1125014+zolotokrylin@users.noreply.github.com>

* Moved "Terminology" section to the top of README.md

* Move Primitive Stream above Composed Stream in TERMINOLOGY.md

* Update STREAM ID definition in terminology

* Fix punctuation

* Update TERMINOLOGY.md with clarification on 'SCHEMA'

* Update TERMINOLOGY.md

* Update TERMINOLOGY.md

Co-authored-by: Vadim <1125014+zolotokrylin@users.noreply.github.com>

* Update project name in TERMINOLOGY.md

* Update terminology definitions for stream relations

The definitions for "CHILD OF" and "PARENT OF" in the terminology document have been updated. They now clearly explain the relationship between streams and include the concept of 'Composed Streams'. The superfluous "COMPOSED OF" definition has been removed, as its meaning was incorporated into the updated definitions.

* Update terminology from 'schema' to 'contract'

The terminology file has been updated, replacing references to 'schemas' with 'contracts'. The updated terms include 'Data Contracts' and 'Upgradeable Contract'. This change better reflects the actual structure and behavior of data in our system.

* Update terminology definitions in documentation

The definitions of 'PRIMITIVE' and 'INDEX' in the terminology documentation have been updated for better clarity. These changes particularly focus on the terms' roles in calculations, helping to ensure a better understanding of their functions in the system.

* Update TERMINOLOGY.md

* Update TERMINOLOGY.md

---------

Co-authored-by: Vadim <1125014+zolotokrylin@users.noreply.github.com>