feat: have more than 1 node supporting TSN (#214)
* Replace gateway setup with PKL config and simplify Docker setup

Removed old gateway configurations and replaced them with a new PKL configuration file. Also simplified the Docker setup by removing the entrypoint.sh file and unnecessary variables, and adding a usage of bash shell.

* Remove dev-gateway Dockerfile and related files

Deleted the dev-gateway Dockerfile, entrypoint.sh, and kgw.base-config.json files. Updated the dev-gateway-compose.yaml to reflect these changes, as they are no longer needed in the current project structure.

* Update environment variable names in gateway-compose.yaml

Environment variables' names in the gateway-compose.yaml file have been fixed. Specifically, 'CHAIN-ID' has been changed to 'CHAIN_ID' and 'SESSION-SECRET' to 'SESSION_SECRET'.

* Refactor TSN and KGW deployment architecture

Extracted AWS CloudFormation resources creation from `cdk_main.go` to relevant submodules. Created new functions and structs for TSN and KGW instances, and added comments on Nitro Enclave and certbot.

* Add tsn-config.dockerfile and update deployment Dockerfiles

This commit adds a new file, tsn-config.dockerfile, to generate a configuration file for a single node or use an externally provided configuration. The Dockerfiles for deployments have also been updated, particularly revising the `ARG` variables and the application of the configuration path. Additionally, the filename `gen-kwil-conf.dockerfile` is changed to `network-conf.dockerfile`.

* Refactor kwil configurations in Dockerfiles and docker-compose files

Improved flexibility and robustness of configuration handling in Dockerfiles and compose files. Introduced a new service in the compose file for configuration handling, enhanced entrypoint.sh checks in Dockerfile for better runtime parameter validation, and removed unnecessary configurations from gateway compose file.

* Fix string syntax at deploy-staging.yaml

* Correct parameter name for CORS configuration

Fixed a typo in the parameter name for CORS allowed origins in the GitHub Action workflow for the staging deployment. It incorrectly used "GW_CORS_ALLOWED_ORIGINS", it has been corrected to "GW_CORS_ALLOW_ORIGINS".

* Update Dockerfile path and argument handling in devnet-compose.yaml

Changed the Dockerfile path to correctly refer to the local deployment. Furthermore, switched from passing arguments during build-time to setting environment variables inside the container. This change simplifies the build process and makes the configuration more consistent across different network nodes.

* Update CDKParams function and add new environment variables configuration

This commit updates the naming case of function parameters in the 'NewCDKParams' function to lowerCamelCase. It also changes the return type of the 'NewCDKParams' function from a pointer to a direct value. Furthermore, a 'GetEnvironmentVariables' function is added which handles fetching and validating environment variables required for the application.

* Add random string to KGWInstance name

The KGWInstance name has been updated to append a random string for redeployment on updates. This is done by introducing a randomString function from the kwil-db/core/utils/random library.

* Update Kwil Gateway startup scripts and refactor environment variables extraction

This commit updates the Kwil Gateway startup scripts, explicitly adding Docker installation and configuration steps while eliminating the dependence on the `tsn-db-app` service. It also refactors the extraction of environment variables to use a utility function, `GetDictFromStruct`, instead of a direct approach using reflection, promoting code reuse. The `reflect` package import has been removed as it was no longer being utilized directly in `kgw_startup_scripts.go`.

* Update TSN cluster creation and configuration

The code now includes new features and improvements to the TSN cluster creation process. It includes the creation of new key pairs, node configuration generation for 'kwil network node', the addition of security group rules for better network communication, and updated startup scripts. It also introduces restructuring and optimization of the import structure for better code readability and maintenance.

* Add TSN config image asset and update output properties

Added a new TSN config image asset in the cdk_main.go file. The public address and TSN nodes count outputs have been updated to use the correct instance IDs and convert the length of tsnCluster.Nodes to a string using strconv.Itoa. Additionally, passed the Params and TSNConfigImageAsset properties to TSNCluster properties.

* Refactor infrastructure libraries and revise import paths

Renamed several directories under 'infra/lib' for clarity, including 'network_utils' to 'kwil-network' and 'gateway_utils' to 'kwil-gateway'. This commit also involves updating related import paths to reflect these changes, improving the consistency and readability of the codebase.

* revert domain -> domain_utils

* Add feature to extract and use peer id from config file

Updated the go.mod and go.sum files to include the BurntSushi/toml package. This package is used to extract a peer id from a configuration file, which is then used to create a new peer connection. Additionally, modified the NewKwilNetworkConfigAssets function to return an array of KwilNetworkConfig objects instead of awss3assets.Asset objects.

* Update method parameters for peer connections and config assets

In the 'tsn_cluster.go' and 'tsn_startup_scripts.go' files, the method parameters for creating peer connections and config assets were updated. More specifically, an additional argument was included in the 'NewPeerConnection' function, and properties referencing 'configAssets' were adjusted. Additionally, the 'GetP2PAddress' function was enhanced with a boolean parameter to provide a more flexible and accurate implementation.

* Refactor code to fetch ChainId from environment variables

The affected code files have been modified to fetch ChainId directly from environment variables instead of the earlier approach of accepting it as a parameter. This eliminates the need for ChainId as a parameter in several functions and client codes, leading to a more streamlined codebase. This change has also been reflected in workflow files where parameters now include the ChainId as environment variable.

* Update ports and change from HTTP to RPC

The various ports used in the compose.yaml file were updated, and the networking code was adjusted to change from HTTP to RPC. This included changes to the security group setup and peer connection logic in the tsn and kwil-network libraries.

* Add kwilteam/kwil-db/core dependency to go.mod

The new dependency, kwilteam/kwil-db/core v0.1.3, has been added to the go.mod file. This change also reflects in the go.sum file, ensuring the correct version of the dependency is used.

* Refactor security group creation in TSN cluster setup

Moved the creation of the security group after the peer connections are set up in tsn_cluster.go. This change allows the initialization of the security group in security_group.go to use the peer connections. This ensures that communication between nodes is allowed using the peer's public IP, not the private IP, fixing connectivity.

* Update path for KWIL_ADMIN_BIN_PATH in deploy-staging.yaml

The KWIL_ADMIN_BIN_PATH in the deploy-staging.yaml file has been updated to make it point to the correct path.

* fix deploy comment