feat: define taxonomy of a composed stream (#264)
* Add taxonomy setting procedures to contract

Added "get_current_version" and "set_taxonomy" procedures to the contract. The "get_current_version" procedure is used to find the current version number from the data in the taxonomies table. The "set_taxonomy" procedure allows the user to set taxonomies in batch, implementing data validity checks and increasing the version number for each new insert. A test section for the new procedures has also been added.

* Update SQL syntax in metadata retrieval

Updated the SQL statement used in the metadata retrieval process. Replaced the 'get_metadata()' function in the 'composed_stream_template.kf' file with a 'SELECT * FROM get_metadata()' to align with SQL syntax.