feat: disable taxonomy of composed contract (#266)
* Add taxonomy setting procedures to contract

Added "get_current_version" and "set_taxonomy" procedures to the contract. The "get_current_version" procedure is used to find the current version number from the data in the taxonomies table. The "set_taxonomy" procedure allows the user to set taxonomies in batch, implementing data validity checks and increasing the version number for each new insert. A test section for the new procedures has also been added.

* Update SQL syntax in metadata retrieval

Updated the SQL statement used in the metadata retrieval process. Replaced the 'get_metadata()' function in the 'composed_stream_template.kf' file with a 'SELECT * FROM get_metadata()' to align with SQL syntax.

* Add 'describe_taxonomies' procedure and related tests

The 'describe_taxonomies' procedure has been added for better taxonomy management. By calling this procedure, detailed taxonomy information can be obtained, such as child stream id, data provider, creation time etc. Accompanying tests have also been included in the commit to ensure the newly added procedure works as expected.

* Refactor 'get_current_version' method to handle disabled taxonomies

The 'get_current_version' method has been updated to include an extra parameter which will allow it to handle disabled taxonomies. Depending on the boolean value of this parameter, the method will either return the version of a taxonomy that is not disabled or just the latest version of any taxonomy.

* Add 'disable_taxonomy' procedure and test

This commit introduces the disable_taxonomy procedure in the composed_stream_template file to allow disabling of taxonomies based on their version. Changes also include an addition of a test scenario in the composed_stream_contract_test.md file to verify the correct functionality of the new procedure.

* resolve conflicts

* resolve conflicts