fix: fetch data through compatible KGW  (#310)
* update IaC to use latest kgw

* Update default BACKENDS port in dev-gateway-compose.yaml

The default port for BACKENDS in the dev-gateway-compose.yaml file was 8080. This commit updates it to 8484 to match current configurations, ensuring proper communication and coordination between services in the development gateway.

* Update gateway version in cdk_main.go

The gateway version key in the cdk_main.go file has been updated. The version has been changed from "kgw-v0.2.0.zip" to "kgw-v0.3.0-beta.1.zip". This update is necessary to keep the deployment infrastructure in sync with the latest release of the gateway package.