feat: deploy with predefined config through GH action (#469)
* Handle string type for numOfNodes context value.

Previously, the numOfNodes context value was assumed to be a float64, leading to potential conversion issues if it was provided as a string. Now, the code checks and properly converts the context value if it's a string, and logs an error if the conversion fails, ensuring more robust handling of context values.

* Remove GENESIS_PATH from environment variables

The GENESIS_PATH environment variable was redundant and unused. This change simplifies the configuration by removing unnecessary variables, making the codebase cleaner and easier to maintain.

* Add inputs and validations for deploy-auto workflow

Enhanced the deploy-auto workflow to allow specifying the number of nodes and subdomain through inputs. Added validation to restrict auto deployments to the 'dev' subdomain only, and ensured parameters are correctly passed and context values are set for deployment flexibility.

* Add NODE_PRIVATE_KEYS secret and refine CDK deploy command

Included NODE_PRIVATE_KEYS secret to the environment variables for deployment. Updated the CDK deploy command to target specific stacks and added a deployment stage context for more precise control. This ensures only necessary stacks are deployed and improves deployment security.

* Add RESTART_HASH docs to control TSN redeployments

* Add GitHub Actions deployment instructions to README

Document new deployment workflows for automated and staging environments using GitHub Actions. Includes setup inputs, AWS deployment details, and references to configuration files.

* Update deploy command to use TSN-From-Config stack

Updated the CDK deploy command in deploy-staging.yaml to target the newly renamed TSN-From-Config stack instead of the previous TSN-DB stack. This includes updating the stack names in the parameters and ensuring the deployment context is set to STAGING.

* Add error handling to genesis.json download in staging deploy

Updated the cURL command to include error handling parameters like `-f`, `--silent`, and `--show-error`. This ensures that any failure in downloading the genesis.json file will provide a clear error message.

* Add genesis file validation and fix deployment script

Added a `validateGenesisFile` function to check the existence and validity of the genesis file in `peer_config_gen.go`. Also, updated the deployment script in `deploy-staging.yaml` to ensure the correct path and parameters are used for the staging environment.

* Update error handling in config and peer config generation

Changed error logs to panic with detailed messages in the config and peer configuration files. This ensures immediate attention to critical issues by including specific file paths and error details in panic statements.