chore: update domain handling and configuration in CDK setup (#903)
* chore: update domain handling and configuration in CDK setup

This commit refines the CDK setup by introducing a centralized domain management system through the new HostedDomain construct. It updates the CertStack and TSN stack to utilize this new domain configuration, ensuring consistent handling of deployment stages and domain names. Additionally, it modifies environment variable references in the configuration files, replacing `KWIL_ADMIN_BIN_PATH` with `KWILD_CLI_PATH` for improved clarity. The README.md is also updated to reflect changes in deployment parameters and commands, enhancing documentation accuracy.

* chore: update environment variable names in deployment workflows

This commit replaces the `KWIL_ADMIN_BIN_PATH` variable with `KWILD_CLI_PATH` in both `deploy-auto.yaml` and `deploy-staging.yaml` workflows. The change aims to improve clarity by providing a more descriptive name for the path to the kwild CLI binary used for configuration generation.

* chore: enhance kwild CLI integration and configuration management

This commit updates the integration with the kwild CLI by refining command usage in the genesis file generation and node key generation processes. It modifies the command parameters for improved clarity and consistency, ensuring that the temporary directory structure is correctly utilized. Additionally, it enhances the peer configuration generation by initializing the configuration directory with the genesis file and overriding settings for peer connections. The changes aim to streamline the configuration management and improve the overall usability of the deployment scripts.