feat: from config stack compatible with upgrade (#912)
* feat: add XFF trust proxy count configuration to gateway and KGW instances

This commit introduces the `XFF_TRUST_PROXY_COUNT` environment variable to the gateway configuration in `gateway-compose.yaml`, allowing for flexible trust proxy settings. Additionally, it updates the `KGWConfig` and `KGWEnvConfig` structures to include the new `XffTrustProxyCount` field, ensuring that the configuration is properly managed across the deployment. The changes enhance the overall security and configurability of the gateway service.

* refactor: update indexer environment configuration to use NODE_RPC_ENDPOINT

This commit modifies the `IndexerEnvConfig` structure in `indexer_startup_scripts.go` to replace the `NodeCometBftEndpoint` field with `NodeRpcEndpoint`, aligning with recent changes in the Docker configuration for indexer services. The update enhances clarity and consistency in the environment setup for the indexer.

* chore: remove KWIL_PG_CONN from indexer environment configuration

* feat: enhance Kwil asset management and indexer startup scripts

This commit introduces several updates to improve the management of Kwil assets and the startup scripts for the indexer. Key changes include the addition of the `IndexerBinaryKey` to `KwilAssetOptions`, enabling the specification of the indexer binary's S3 location. The `BuildKwilAssets` function is updated to include the binary asset for the indexer, ensuring it is packaged alongside the other assets. Additionally, the `IndexerEnvConfig` structure is modified to include the `IndexerBinaryPath`, and the startup scripts are enhanced to handle the unzipping and execution of the indexer binary. These changes aim to streamline the deployment process and improve the overall functionality of the indexer service.

* feat: enhance API Gateway request path handling

This commit updates the API Gateway integration in `api_gateway.go` to include an additional parameter mapping for the request path. The new `OverwritePath` method is added to improve routing capabilities, ensuring that the API can handle requests more flexibly. These changes enhance the overall functionality and maintainability of the API Gateway configuration.

* feat: update environment variable management and enhance node configuration

This commit introduces the `IncludeObserver` field to the `MainEnvironmentVariables` structure, improving the management of environment variables. It also refactors the `KwilNetworkConfigAssetsFromNumberOfNodes` function to handle private keys more effectively, ensuring that the number of nodes matches the provided keys. Additionally, the `TnAutoStack` and `TnFromConfigStack` files are updated to utilize the new environment variable, enhancing the overall configuration flexibility and clarity in the deployment process.

* feat: refactor environment variable annotations to fix behavior

* feat: add IncludeRole field to AttachInitDataToLaunchTemplate for enhanced instance configuration

* feat: add genesis file chain ID verification to Kwil network configuration

This commit introduces a new function, `verifyGenesisChainID`, which reads and verifies the `chain_id` from the provided genesis file against the expected value. If the verification fails, it panics with an appropriate error message. This enhancement ensures that the genesis file used in the network configuration is valid and matches the expected chain ID, improving the robustness of the deployment process.

* fix: update PROVIDER environment variable in Taskfile.yml for local development

This commit modifies the `PROVIDER` environment variable in the `Taskfile.yml` to point to `http://localhost:8484`, enhancing the configuration for local development and ensuring that migration tasks utilize the correct endpoint.